Extend base endpoint instead of implement
