--- common/base_endpoint.d.ts
@@ -1,19 +1,0 @@
-import { Knex } from "knex";
-import { RedisClientType } from "redis";
-import winston from "winston";
-
-import { ENV } from "../app";
-
-declare class IBaseEndpoint {
-  public allowNames: Array<string>;
-
-  constructor(
-    public db: Knex,
-    public redisClient: RedisClientType,
-    public logger: winston.Logger,
-    public config: ENV
-  );
-
-  async callEndpoint(name: string, args: object, user: TUser | undefined): Promise<object>;
-  async validate<EType>(schema: unknown, args: EType): Promise<EType>;
-}
