+++ endpoints/permission/permission_endpoint.ts
@@ -1,12 +1,10 @@
-import Joi from "joi";
 import { Knex } from "knex";
 import { RedisClientType } from "redis";
 import winston from "winston";
 
 import { ENV } from "../../app";
-import { IBaseEndpoint } from "../../common/base_endpoint";
+import BaseEndpoint from "../../common/base_endpoint";
 import RequestError from "../../common/request_error";
-import { EndpointThisType } from "../../common/types";
 import CachingPermissionModel from "../../db/models/caching/caching_permission";
 import CachingUserModel from "../../db/models/caching/caching_user";
 import PermissionModel, { TPermissions } from "../../db/models/permission";
@@ -15,7 +13,7 @@ import * as type from "./types";
 
 type CallEndpointReturnType = {} | TPermissions | {success: true};
 
-export default class PermissionEndpoint implements IBaseEndpoint {
+export default class PermissionEndpoint extends BaseEndpoint<type.PermissionRequestArgs, CallEndpointReturnType> {
   allowNames: Array<string> = ["view", "grant", "rescind"];
   permissionModel: PermissionModel | CachingPermissionModel;
   userModel: UserModel | CachingUserModel;
@@ -26,6 +24,8 @@ export default class PermissionEndpoint implements IBaseEndpoint {
     public logger: winston.Logger,
     public config: ENV
   ) {
+    super(db, redisClient, logger, config, "permission");
+
     const REDIS_REQUIRED = this.config.get("REDIS_REQUIRED").required().asBool();
     const UserModelType = REDIS_REQUIRED ? CachingUserModel : UserModel;
     const PermissionModelType = REDIS_REQUIRED ? CachingPermissionModel : PermissionModel;
@@ -70,27 +70,4 @@ export default class PermissionEndpoint implements IBaseEndpoint {
     return { success: true };
   }
   // >>> Rescind >>>
-
-  async callEndpoint(
-    this: EndpointThisType<PermissionEndpoint, type.PermissionRequestArgs, Promise<CallEndpointReturnType>>,
-    name: string, args: type.PermissionRequestArgs, user: TUser | undefined
-  ): Promise<CallEndpointReturnType> {
-    const permissionIncludes = this.allowNames.includes(name);
-    if (!permissionIncludes) {
-      throw new RequestError("EndpointNotFound", `Endpoint permission/${name} does not exist`, 404);
-    }
-
-    const result: CallEndpointReturnType = await this[name](args, user);
-
-    return result;
-  }
-
-  async validate<EType>(schema: Joi.ObjectSchema, args: EType): Promise<EType> {
-    const { error, value } = schema.validate(args);
-    if (error) {
-      throw new RequestError("ValidationError", error.message, 400);
-    }
-
-    return value as EType;
-  }
 }
