+++ endpoints/vote/vote_endpoint.ts
@@ -1,12 +1,10 @@
-import Joi from "joi";
 import { Knex } from "knex";
 import { RedisClientType } from "redis";
 import winston from "winston";
 
 import { ENV } from "../../app";
-import { IBaseEndpoint } from "../../common/base_endpoint";
+import BaseEndpoint from "../../common/base_endpoint";
 import RequestError from "../../common/request_error";
-import { EndpointThisType } from "../../common/types";
 import CachingPermissionModel from "../../db/models/caching/caching_permission";
 import CachingPostModel from "../../db/models/caching/caching_post";
 import CachingUserModel from "../../db/models/caching/caching_user";
@@ -18,7 +16,7 @@ import * as type from "./types";
 
 type CallEndpointReturnType = { success: true } | { count: number, voteType: Vote };
 
-export default class VoteEndpoint implements IBaseEndpoint {
+export default class VoteEndpoint extends BaseEndpoint<type.VoteRequestArgs, CallEndpointReturnType> {
   public allowNames: string[] = [
     "vote", "voteCount"
   ];
@@ -33,6 +31,8 @@ export default class VoteEndpoint implements IBaseEndpoint {
     public logger: winston.Logger,
     public config: ENV
   ) {
+    super(db, redisClient, logger, config, "vote");
+
     const REDIS_REQUIRED = this.config.get("REDIS_REQUIRED").required().asBool();
     const UserModelType = REDIS_REQUIRED ? CachingUserModel : UserModel;
     const PermissionModelType = REDIS_REQUIRED ? CachingPermissionModel : PermissionModel;
@@ -80,27 +80,4 @@ export default class VoteEndpoint implements IBaseEndpoint {
     return { count: count, voteType: args.voteType };
   }
   // <<< Vote count <<<
-
-  async callEndpoint(
-    this: EndpointThisType<VoteEndpoint, type.VoteRequestArgs, Promise<CallEndpointReturnType>>,
-    name: string, args: type.VoteRequestArgs, user: TUser | undefined
-  ): Promise<CallEndpointReturnType> {
-    const userIncludes = this.allowNames.includes(name);
-    if (!userIncludes) {
-      throw new RequestError("EndpointNotFound", `Endpoint vote/${name} does not exist`, 404);
-    }
-
-    const result: CallEndpointReturnType = await this[name](args, user);
-
-    return result;
-  }
-
-  async validate<EType>(schema: Joi.ObjectSchema, args: EType): Promise<EType> {
-    const { error, value } = schema.validate(args);
-    if (error) {
-      throw new RequestError("ValidationError", error.message, 400);
-    }
-
-    return value as EType;
-  }
 }
